+++ test/Splunk.OpenTelemetry.AutoInstrumentation.IntegrationTests/BuildTests.DistributionStructure_osx.verified.txt
@@ -46,7 +46,7 @@
   /net/OpenTelemetry.Shims.OpenTracing.dll,
   /net/OpenTelemetry.dll,
   /net/OpenTracing.dll,
-  /net/Splunk.OpenTelemetry.AutoInstrumentation.Plugin.dll,
+  /net/Splunk.OpenTelemetry.AutoInstrumentation.dll,
   /net/System.Configuration.ConfigurationManager.dll,
   /net/System.Diagnostics.PerformanceCounter.dll,
   /net/System.Drawing.Common.dll,
@@ -62,7 +62,7 @@
   /net/runtimes/win/lib/netcoreapp3.0/System.Drawing.Common.dll,
   /net/runtimes/win/lib/netcoreapp3.0/System.Windows.Extensions.dll,
   /net/runtimes/win/lib/netstandard2.0/System.Security.Cryptography.ProtectedData.dll,
-  /netfx/Splunk.OpenTelemetry.AutoInstrumentation.Plugin.dll,
+  /netfx/Splunk.OpenTelemetry.AutoInstrumentation.dll,
   /store/x64/net6.0/artifact.xml,
   /store/x64/net6.0/dnsclient/1.4.0/lib/netstandard2.1/DnsClient.dll,
   /store/x64/net6.0/microsoft.extensions.configuration.abstractions/6.0.0/lib/netstandard2.0/Microsoft.Extensions.Configuration.Abstractions.dll,
