+++ test/Splunk.OpenTelemetry.AutoInstrumentation.IntegrationTests/BuildTests.DistributionStructure_windows.verified.txt
@@ -45,7 +45,7 @@
   \net\OpenTelemetry.Shims.OpenTracing.dll,
   \net\OpenTelemetry.dll,
   \net\OpenTracing.dll,
-  \net\Splunk.OpenTelemetry.AutoInstrumentation.Plugin.dll,
+  \net\Splunk.OpenTelemetry.AutoInstrumentation.dll,
   \net\System.Configuration.ConfigurationManager.dll,
   \net\System.Diagnostics.PerformanceCounter.dll,
   \net\System.Drawing.Common.dll,
@@ -98,7 +98,7 @@
   \netfx\OpenTelemetry.Shims.OpenTracing.dll,
   \netfx\OpenTelemetry.dll,
   \netfx\OpenTracing.dll,
-  \netfx\Splunk.OpenTelemetry.AutoInstrumentation.Plugin.dll,
+  \netfx\Splunk.OpenTelemetry.AutoInstrumentation.dll,
   \netfx\System.AppContext.dll,
   \netfx\System.Buffers.dll,
   \netfx\System.Collections.Concurrent.dll,
