+++ test/Splunk.OpenTelemetry.AutoInstrumentation.Tests/xunit.runner.json
